+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For details about compatibility between different releases, see the **Commitment
 ### Added
 
 - Add recvTime field to the decodeUplink input in payload formatters
+- Add the latest battery percentage of the end device in the `ApplicationUplink` message.
 
 ### Changed